nvim: switch from ultisnips to luasnip
Not all snippets have been ported over, so I'm keeping the ultisnips
directory where it is. I've added the ones I use most often in Go, and
think it's best to just lazily add what I use in various filetypes.

I experimented with loading pre-made snippets but wasn't happy with
pulling these from the plugin. Eventually, it may make sense to clone
the repo into the place the ultisnips directory is and update it to get
it the way I want it to be. As is, I probably wouldn't use or know to
use most of those snippets.