nvim: disable vim-go features redundant with lsp
this has us defer entirely to the lsp client for things like formatting.
a bit of extra work was needed to recreate g:go_imports_autosave,
because lsp treats importing and formatting as separate things and
rather than just automatically organizing imports on save, provides a
codeAction for doing so.

this should result in faster saves and _might_ get rid of that annoying
thing where nvim-compe's json shows up in the quickfix buffer from
vim-go.