nvim: remove vim-go and update treesitter config
I've pretty much only been using for `if` select mappings, so I've
ripped it out and replaced it with equivalent treesitter-textobjects
mappings.

Added lots of new treesitter mappings for navigation and selection.

This also updates treesitter config for the neovim 0.7 release and adds
treesitter-playground, as well as vim-matchup, which is is a replacement
for % mappings which is treesitter aware.