nvim: switch treesitter plugin to nvim-treesitter.withAllGrammers
I think the old syntax was wrong for including all grammars, or had been
replaced. At the same time, some kind of regression in vim-matchup is
not playing nicely with the new treesitter grammars and is making things
very slow to respond, so I'm pinning that version for now and I'll check
up on it periodically.