nvim: update vim-matchup and use fewer treesitter grammars
I was pinning vim-matcup's version to something old to get around
performance issues I was noticing with a large number of grammars.

With the nvim 0.9 release, this version is using an out of date API that
results in ugly warnings.

I opened an issue with vim-matchup, but until there's a fix I'll just
use fewer grammars.

See: https://github.com/andymass/vim-matchup/issues/283