nvim: cleanup and translate init.vim to init.lua
Went through some plugins and configs to convert any lingering vimscript
config to lua. The old init.vim is still around in case I find I need
some of what I omitted or translated something incorrectly.

Did a bit of cleanup while I was at it. Most notable changes are:

* Dropped sensible.vim. Most of the defaults it set were already set by
  neovim and I could translate a couple of others into my init.lua.
* Got rid of base64 and scratch plugins. I rarely used these and can
  bring them back if I ever miss them.

  base64 should be relatively easy to replace with a quick script to
  shell out to coreutils.base64.
* Added a couple of new mappings and removed some that I was never using
* vim-matchup is no longer pinned to a revision because the performance
  problems I was seeing were addressed:
  https://github.com/andymass/vim-matchup/issues/283
* No longer load unmanaged quickload.vim from init.vim or offer a lua
  equivalent. I wasn't using this.

I discovered a couple of places where some plugin/plugin configs depend
on others in ways that are not well documented and not forced by nix.
The biggest offender is nvim-cmp, which depends on all of its plugins
and on luasnip, and copying just my cmp config won't work unless you
also use luasnip and bring over dependent plugins. It'd be nice to
figure out a way to make this dependency explicit.