nvim: configure lualine to show jj change id
Quick and dirty change to show the current jj change ID instead of git
commit/branch. The config checks whether the working directory is in a
jj repo on init to know what to show in that section.

This means restarting vim if we initialize a jj repo after startup. Not
ideal, but it works.

Another minor quirk is that this still uses lualine's built-in git diff.
The git diff may not always correspond to diffs in the current jj
change; in the future, it may be worth seeing if there's a clean/easy
way to get the info from jj, and/or it may make more sense to just show
whether or not the current buffer's file shows up in 'jj status'.