nvim: add bat as explicit path dependency
While debugging fzf issues, I learned that fzf will subprocess bat when
it's available to color the contents of the preview window.

The way it worked before was a happy accident born of the fact that I
had bat installed anyway, so now things shouldn't break if I ever remove
it from home.nix.