jj: add watchman for nixpkgs repo
The nixpkgs repo is really large and the [docs] suggest that setting
core.fsmonitor = "watchman" prevents jj from rescanning the entire
working copy.

This feels like it made a difference, even without seting
core.watchman.register_snapshot_trigger to create snapshots on
filesystem changes.

It remains to be seen if there's value in setting this globally, or only
for very large repos. For now, I only updated that repo's config.

In the future, it might be a good idea to use makeWrapper to get
watchman out of my $PATH and explicitly add it to jj's.

[docs]: https://github.com/jj-vcs/jj/blob/31afda99d4f97be7d41e5ea77f5f59e0ddab1e08/docs/config.md#filesystem-monitor