nix: patch tailscale on darwin systems
A commit in nixpkgs introduces
some wrapping that breaks a hack I had in place. This is kind of a quick
and dirty way to undo that.

The downside is that lsof is no longer explicitly in the $PATH of the
underlying executable, but this doesn't matter for my purposes.

I updated the nixpkgs input was updated to point to my local checkout of
nixpkgs and found the offending commit with git bisect and the following
script:

	#!/bin/sh
	(cd "$HOME/.config"; nix flake update nixpkgs)
	nix run nix-darwin -- -v -L switch --flake "path:$HOME/.config"
	tailscale status

See: https://github.com/NixOS/nixpkgs/pull/375407