nix: break go overlay into parts
The first part continues to live in mkSystem - largely unnecessary but
helpful to get ahead on go versions where there's an update that hasn't
made its way to nixpkgs yet.

The original purpose of the overlay was to patch fsync in darwin;
this has been moved to only work machines. This also re-adds all old
patches instead of clobbering patches and leaving only the fsync patch.