nix: add nvim flake output
Experimenting with exposing nvim as a flake output so I can share a
fully-configured nvim via nix flake:

	nix run github:cixel/config#nvim

This is a good way to showcase how Nix works to contain self-contained
packages, and it could be sorta useful to be able to drop into an editor
I'm used to wherever Nix is available.

There are some gaps in language server support - I currently just expect
gopls and rust-analyzer's deps to exist and don't bring them in
explicitly. In the future, maybe I can add a default installation for
those to use as fallbacks.